move the test for main back to /unit tests
